[MacOS] Share ViewRenderer with iOS
